Setted difference between app and server
