+++ package.json
@@ -7,7 +7,7 @@
         "@types/jsonwebtoken": "^8.5.6",
         "@types/morgan": "^1.9.3",
         "@types/mysql": "^2.15.19",
-        "@types/node": "^14.11.2",
+        "@types/node": "^14.18.5",
         "@types/swagger-ui-express": "^4.1.3",
         "gts": "^3.1.0",
         "tslint": "^6.1.3",
